+++ internal/util/iputil/util_test.go
@@ -38,7 +38,7 @@ func TestGetIpAndPort(t *testing.T) {
 
 func TestGetIpAndInvalidPort(t *testing.T) {
 	testAddress := "121.1.23.3:aa"
-	if ip, port := GetIPAndPort(testAddress); ip != "121.1.23.3" || port != int32(defaultPort) {
+	if ip, port := GetIPAndPort(testAddress); ip != "121.1.23.3" || port != defaultPort {
 		t.Fatal("GetIPAndPort failed.")
 	}
 }
